Find Maximum Float Number

+++ FindMaximumNumber.java
@@ -2,8 +2,8 @@ package generics;
 
 public class FindMaximumNumber {
 
-	public static Integer findMaximumNumber(Integer x, Integer y, Integer z) {
-		Integer max = x;
+	public static Float findMaximumNumber(Float x, Float y, Float z) {
+		Float max = x;
 		if (y.compareTo(max) > 0) {
 			max = y;
 		}
